:sparkles: Basic server semver version check
If the semver from the latest on GitHub is different than the local build, it will present a message to update your server. I did not make this a toast message,  at least not yet. I figure that it might be neat to allow server updates to be configured as a notification once the notifier feature is in place