:bug: Fix OPDS path detection regression
The recent change to split the main OPDS router into sub-routers for each supported version caused some issues in the auth middleware. Specifically, the `/opds` prefix was missing which caused the conditional check which allows basic auth for OPDS-specific routes. This is expected behavior with the changes, since accessing the request parts `uri` in an extension is scoped to the router, which after nesting multiple routers under the main opds one means the top-level is v1.2