:adhesive_bandage: Fix path filter on push (#388)
I noticed that while the path filters work as expected for PR events, once merged the base which it uses to diff against is main, which means there will almost always be code-related changes and so the builds don't get skipped. By specifying the base, I'm hoping this gets fixed. The docs state this arg is ignored for PR events, so should have no effect there